pyproject.toml
[tool.poetry]
name = "video-service"
version = "0.1.0"
description = "Video workflow microservice"
authors = ["Codex"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.24.0"
kafka-python = "^2.0.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.3"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
